ConsultoSec/backend/consultas/serializers.py
from rest_framework import serializers
from .models import Consulta

class ConsultaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Consulta
        fields = '__all__'  # Esto expone todos los campos del modelo