ConsultoSec/backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
